+++ debug.F90
@@ -2,8 +2,8 @@
 module debug
 
 ! debuging flags
-integer :: idebug = 0
-integer :: iprint = 0
-integer :: iinfo  = 0
+integer(kind=4) :: idebug = 0
+integer(kind=4) :: iprint = 0
+integer(kind=4) :: iinfo  = 0
 
 end module
